# server/docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: OpsPilot API
  version: 0.7.0-dev
  description: OpsPilot auth, RBAC, Agent, Host, script, task execution, logs, scheduling, webhooks, metrics, alerts, and notifications API.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/v1/webhooks/trigger/{token}:
    post:
      security: []
      summary: Trigger webhook source
      description: Requires a valid source token, HMAC signature, `X-OpsPilot-Timestamp`, and `X-OpsPilot-Nonce`.
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
        - name: X-OpsPilot-Signature
          in: header
          schema: { type: string }
        - name: X-Hub-Signature-256
          in: header
          schema: { type: string }
        - name: X-OpsPilot-Timestamp
          in: header
          schema: { type: string, example: "1735689600" }
        - name: X-OpsPilot-Nonce
          in: header
          schema: { type: string, example: "nonce-123" }
        - name: X-Delivery-Id
          in: header
          schema: { type: string }
        - name: X-Event-Type
          in: header
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200": { description: Trigger result with matched task runs }
  /api/v1/webhooks/sources:
    get:
      summary: List webhook sources
      description: Requires `webhook:read`.
      responses:
        "200": { description: Source list }
    post:
      summary: Create webhook source
      description: Requires `webhook:manage`. Returns trigger token and signing secret once.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookSourcePayload"
      responses:
        "200": { description: Source detail with one-time trigger token and signing secret }
  /api/v1/webhooks/sources/{id}/rotate-secret:
    post:
      summary: Rotate webhook source signing secret and optional trigger token
      description: Requires `webhook:manage`.
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                rotateToken: { type: boolean, default: false }
      responses:
        "200": { description: Source detail with one-time signing secret and optional trigger token }
  /api/v1/webhooks/sources/{id}/pause:
    post:
      summary: Pause webhook source
      description: Requires `webhook:manage`.
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200": { description: OK }
  /api/v1/webhooks/sources/{id}/resume:
    post:
      summary: Resume webhook source
      description: Requires `webhook:manage`.
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200": { description: OK }
  /api/v1/webhooks/sources/{id}/disable:
    post:
      summary: Disable webhook source
      description: Requires `webhook:manage`.
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200": { description: OK }
  /api/v1/webhooks/rules:
    get:
      summary: List webhook rules
      description: Requires `webhook:read`.
      responses:
        "200": { description: Rule list }
    post:
      summary: Create webhook rule
      description: Requires `webhook:manage`.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookRulePayload"
      responses:
        "200": { description: Rule summary }
  /api/v1/webhooks/rules/{id}:
    put:
      summary: Update webhook rule
      description: Requires `webhook:manage`.
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                eventType: { type: string }
                matcher:
                  type: object
                  properties:
                    conditions:
                      type: array
                      items:
                        type: object
                        required: [type, value]
                        properties:
                          type: { type: string, enum: [header_equals, payload_equals, payload_contains, event_type_equals, ref_equals, branch_equals] }
                          key: { type: string, example: X-GitHub-Event }
                          path: { type: string, example: repository.name }
                          value: { type: string, example: push }
      responses:
        "200": { description: Updated rule summary }
  /api/v1/webhooks/rules/{id}/pause:
    post:
      summary: Pause webhook rule
      description: Requires `webhook:manage`.
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200": { description: OK }
  /api/v1/webhooks/rules/{id}/resume:
    post:
      summary: Resume webhook rule
      description: Requires `webhook:manage`.
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200": { description: OK }
  /api/v1/webhooks/rules/{id}/disable:
    post:
      summary: Disable webhook rule
      description: Requires `webhook:manage`.
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200": { description: OK }
  /api/v1/webhooks/matcher/simulate:
    post:
      summary: Simulate webhook matcher evaluation
      description: Requires `webhook:manage`.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                ruleId: { type: string }
                eventType: { type: string }
                headers:
                  type: object
                  additionalProperties: { type: string }
                payload:
                  type: string
                matcher:
                  type: object
                  properties:
                    conditions:
                      type: array
                      items:
                        type: object
                        required: [type, value]
                        properties:
                          type: { type: string, enum: [header_equals, payload_equals, payload_contains, event_type_equals, ref_equals, branch_equals] }
                          key: { type: string }
                          path: { type: string }
                          value: { type: string }
      responses:
        "200": { description: Matcher simulation result }
  /api/v1/webhook-events:
    get:
      summary: List webhook events
      description: Requires `webhook:read`.
      parameters:
        - name: sourceId
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string, enum: [received, ignored, triggered, failed, rejected] }
        - name: deliveryId
          in: query
          schema: { type: string }
        - name: receivedFrom
          in: query
          schema: { type: string, example: "2026-05-01 00:00:00" }
        - name: receivedTo
          in: query
          schema: { type: string, example: "2026-05-08 23:59:59" }
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
      responses:
        "200": { description: Paged webhook event list }
  /api/v1/webhooks/events/{id}:
    get:
      summary: Get webhook event detail
      description: Requires `webhook:read`.
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200": { description: Webhook event detail with headers, payload, and match results }
  /api/v1/webhooks/events/{id}/replay:
    post:
      summary: Replay webhook event
      description: Requires `webhook:manage`.
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                simulateOnly: { type: boolean, default: false }
                idempotencyKey: { type: string }
      responses:
        "200": { description: Trigger result with replay status and triggered runs }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    agentTokenAuth:
      type: http
      scheme: bearer
      description: Agent Token, not user JWT.
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: { type: string }
...
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    pageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  schemas:
...
    WebhookSourcePayload:
      type: object
      required: [name]
      properties:
        name: { type: string }
        sourceType: { type: string, enum: [github, gitee, gitlab, custom], default: custom }
    WebhookRulePayload:
      type: object
      required: [sourceId, name]
      properties:
        sourceId: { type: string }
        targetType: { type: string, enum: [task, workflow], default: task }
        taskId: { type: string }
        workflowId: { type: string }
        name: { type: string }
        eventType: { type: string }
        matcher:
          type: object
          properties:
            conditions:
              type: array
              items:
                type: object
                required: [type, value]
                properties:
                  type: { type: string, enum: [header_equals, payload_equals, payload_contains, event_type_equals, ref_equals, branch_equals] }
                  key: { type: string, example: X-GitHub-Event }
                  path: { type: string, example: repository.name }
                  value: { type: string, example: push }
...